+++ scout/server/links.py
@@ -63,6 +63,15 @@ def add_gene_links(gene_obj, build=37):
     gene_obj["cbioportal_link"] = cbioportal_gene(hgnc_symbol)
     gene_obj["civic_link"] = civic_gene(hgnc_symbol)
     gene_obj["iarctp53_link"] = iarctp53(hgnc_symbol)
+    gene_obj["stripy_link"] = stripy_gene(hgnc_symbol)
+
+
+def stripy_gene(hgnc_symbol):
+    link = "//https://stripy.org/database/{}"
+
+    if not hgnc_symbol:
+        return None
+    return link.format(hgnc_symbol)
 
 
 def civic_gene(hgnc_symbol):