+++ CHANGELOG.md
@@ -11,6 +11,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Add a toggle for melter rerun monitoring of cases
 - Add a config option to show the rerun monitoring toggle
 - Add a cli option to export cases with rerun monitoring enabled
+- Add a link to STRipy for STR variants
 ### Changed
 - Slightly smaller and improved layout of content in case PDF report
 - Relabel more cancer variant pages somatic for navigation